Fixed the exponential distribution and added it
(+) Let the max of λe^(-λx) run out to x = 10, which acts as a working
approximation of ∞ while not eating up processing time. RVS and
accept/reject run at about 1e6 iterations, but that’s ~3s computation
time on my laptop, so not too bad. Tightened the error margin to 2σ to
make sure the x=10 approximation worked.

(Δ) Normed the parabolic approximation. Noticed in class that my
distribution was in the faulty bracket. Not sure why unless that copy
was an older edition, which would make sense